backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
fastapi-mail==1.4.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
